+++ src/arvuvdevice.c
@@ -120,7 +120,7 @@ arv_uv_device_create_stream (ArvDevice *device, ArvStreamCallback callback, void
 }
 
 static gboolean
-_read_memory (ArvUvDevice *uv_device, guint32 address, guint32 size, void *buffer, GError **error)
+_read_memory (ArvUvDevice *uv_device, guint64 address, guint32 size, void *buffer, GError **error)
 {
 	ArvUvcpPacket *packet;
 	void *read_packet;
@@ -213,7 +213,7 @@ _read_memory (ArvUvDevice *uv_device, guint32 address, guint32 size, void *buffe
 }
 
 static gboolean
-arv_uv_device_read_memory (ArvDevice *device, guint32 address, guint32 size, void *buffer, GError **error)
+arv_uv_device_read_memory (ArvDevice *device, guint64 address, guint32 size, void *buffer, GError **error)
 {
 	ArvUvDevice *uv_device = ARV_UV_DEVICE (device);
 	int i;
@@ -234,7 +234,7 @@ arv_uv_device_read_memory (ArvDevice *device, guint32 address, guint32 size, voi
 }
 
 static gboolean
-_write_memory (ArvUvDevice *uv_device, guint32 address, guint32 size, void *buffer, GError **error)
+_write_memory (ArvUvDevice *uv_device, guint64 address, guint32 size, void *buffer, GError **error)
 {
 	ArvUvcpPacket *packet;
 	void *read_packet;
@@ -326,7 +326,7 @@ _write_memory (ArvUvDevice *uv_device, guint32 address, guint32 size, void *buff
 }
 
 static gboolean
-arv_uv_device_write_memory (ArvDevice *device, guint32 address, guint32 size, void *buffer, GError **error)
+arv_uv_device_write_memory (ArvDevice *device, guint64 address, guint32 size, void *buffer, GError **error)
 {
 	ArvUvDevice *uv_device = ARV_UV_DEVICE (device);
 	int i;
@@ -347,13 +347,13 @@ arv_uv_device_write_memory (ArvDevice *device, guint32 address, guint32 size, vo
 }
 
 static gboolean
-arv_uv_device_read_register (ArvDevice *device, guint32 address, guint32 *value, GError **error)
+arv_uv_device_read_register (ArvDevice *device, guint64 address, guint32 *value, GError **error)
 {
 	return arv_uv_device_read_memory (device, address, sizeof (guint32), value, error);
 }
 
 static gboolean
-arv_uv_device_write_register (ArvDevice *device, guint32 address, guint32 value, GError **error)
+arv_uv_device_write_register (ArvDevice *device, guint64 address, guint32 value, GError **error)
 {
 	return arv_uv_device_write_memory (device, address, sizeof (guint32), &value, error);
 }
